A hasty hacky change to use doubles instead of floats. Change the scales to no longer need hax factors - although some need to be multiplied out still. Add a time scale factor, which needs to be in the millions for visible motion in this setup.

+++ app.h
@@ -36,7 +36,7 @@ protected:
   
   virtual void HandleEvent(sf::Event const& _event) = 0;
   
-  virtual void UpdateState(float const _dt) = 0;
+  virtual void UpdateState(double const _dt) = 0;
   
   virtual void RenderState() = 0;
